Reorganization + Adding Content, Created FB Page
I did several things. I merged several pages that seemed redundant
(Mission is already covered in About Us). I added a variety of
resources from the YURA resources page to the AURA site as well, for
further visibility. Moved around some content that was clustered. (Ex:
split out part of the About Us Page and put it in the Affiliated
Organizations Page). Restructured the Conference Page. Also, I created
a Facebook page for AURA, and linked it to the AURA site.